+++ src/main/resources/application.properties
@@ -8,9 +8,10 @@ spring.jpa.hibernate.ddlAuto=update
 spring.h2.console.settings.web-allow-others=true
 
 app.secret=aM7l+SXuEgqITZT5JI9uZK9IRcuIMk9ww2K1udnSk1U=
+#1 час
 app.refreshTokenExpTime=3600000
+#10 минут
 app.accessTokenExpTime=600000
 app.issuer=AUTH_SERVER
 
 server.port=8080
-```
